ampute.default.patterns {mice}	R Documentation

Default patterns in ampute

Description

This function creates a default pattern matrix for the multivariate amputation function ampute().

Usage

ampute.default.patterns(n)

Arguments

n
A scalar specifying the number of variables in the data.

Value

A square matrix of size n where 0 indicates a variable

Author(s)

Rianne Schouten, 2016

See Also

ampute, md.pattern

[Package mice version 3.14.0 Index]