ampute.default.weights {mice}	R Documentation

Default weights in ampute

Description

Defines the default weights matrix for the multivariate amputation function ampute.

Usage

ampute.default.weights(patterns, mech)

Arguments

patterns
A matrix of size #patterns by #variables where 0 indicates a variable should have missing values and 1 indicates a variable should remain complete. Could be the result of ampute.default.patterns.

mech
A string specifying the missingness mechanism.

Value

A matrix of size #patterns by #variables containing the weights that will be used to calculate the weighted sum scores. Equal weights are given to all variables. When mechanism is MAR, variables that will be amputed will be weighted with 0. If it is MNAR, variables that will be observed will be weighted with 0. If mechanism is MCAR, the weights matrix will not be used. A default MAR matrix will be returned.

Author(s)

Rianne Schouten, 2016

See Also

ampute, ampute.default.patterns

[Package mice version 3.14.0 Index]