Appends specified break to the data

A custom function to insert rows in long data with new pseudo-observations that are being done on the specified break ages. There should be a column called first in data with logical data that codes whether the current row is the first for subject id. Furthermore, the function assumes that columns age, occ, hgt.z, wgt.z and bmi.z are available. This function is used on the tbc data in FIMD chapter 9. Check that out to see it in action.

Usage

appendbreak(data, brk, warp.model = warp.model, id = NULL, typ = "pred")

Arguments

data

A data frame in the long long format

brk

A vector of break ages

warp.model

A time warping model

id

The subject identifier

typ

Label to signal that this is a newly added observation

Value

A long data frame with additional rows for the break ages