Create a mira object from repeated analyses

The as.mira() function takes the results of repeated complete-data analysis stored as a list, and turns it into a mira object that can be pooled.

Usage

as.mira(fitlist)

Arguments

fitlist

A list containing $m$ fitted analysis objects

Value

An S3 object of class mira.

See also

Author

Stef van Buuren