Extends a formula with predictors

Extends a formula with predictors

extend.formula(formula = ~0, predictors = NULL, auxiliary = TRUE,
  include.intercept = FALSE, ...)

Arguments

formula
A formula. If it is not a formula, the formula is internally reset to ~0.

predictors
A character vector of variable names.

auxiliary
A logical that indicates whether the variables listed in predictors should be added to the formula as main effects. The default is TRUE.

include.intercept
A logical that indicated whether the intercept should be included in the result.

Value

A formula