fdd {mice}	R Documentation

SE Fireworks disaster data

Description

Multiple outcomes of a randomized study to reduce post-traumatic stress.

Format

fdd is a data frame with 52 rows and 65 columns:

id

Client number

trt

Treatment (E=EMDR, C=CBT)

pp

Per protocol (Y/N)

trtp

Number of parental treatments

sex

Sex: M/F

etn

Ethnicity: NL/OTHER

age

Age (years)

trauma

Trauma count (1-5)

prop1

PROPS total score T1

prop2

PROPS total score T2

prop3

PROPS total score T3

crop1

CROPS total score T1

crop2

CROPS total score T2

crop3

CROPS total score T3

masc1

MASC score T1

masc2

MASC score T2

masc3

MASC score T3

cbcl1

CBCL T1

cbcl3

CBCL T3

prs1

PRS total score T1

prs2

PRS total score T2

prs3

PRS total score T3

ypa1

PTSD-RI B intrusive recollection parent T1

ypb1

PTSD-RI C avoidant/numbing parent T1

ypc1

PTSD-RI D hyper-arousal parent T1

yp1

PTSD-RI B+C+D parent T1

ypa2

PTSD-RI B intrusive recollection parent T2

ypb2

PTSD-RI C avoidant/numbing parent T2

ypc2

PTSD-RI D hyper-arousal parent T2

yp2

PTSD-RI B+C+D parent T1

ypa3

PTSD-RI B intrusive recollection parent T3

ypb3

PTSD-RI C avoidant/numbing parent T3

ypc3

PTSD-RI D hyper-arousal parent T3

yp3

PTSD-RI B+C+D parent T3

yca1

PTSD-RI B intrusive recollection child T1

ycb1

PTSD-RI C avoidant/numbing child T1

ycc1

PTSD-RI D hyper-arousal child T1

yc1

PTSD-RI B+C+D child T1

yca2

PTSD-RI B intrusive recollection child T2

ycb2

PTSD-RI C avoidant/numbing child T2

ycc2

PTSD-RI D hyper-arousal child T2

yc2

PTSD-RI B+C+D child T2

yca3

PTSD-RI B intrusive recollection child T3

ycb3

PTSD-RI C avoidant/numbing child T3

ycc3

PTSD-RI D hyper-arousal child T3

yc3

PTSD-RI B+C+D child T3

ypf1

PTSD-RI parent full T1

ypf2

PTSD-RI parent full T2

ypf3

PTSD-RI parent full T3

ypp1

PTSD parent partial T1

ypp2

PTSD parent partial T2

ypp3

PTSD parent partial T3

ycf1

PTSD child full T1

ycf2

PTSD child full T2

ycf3

PTSD child full T3

ycp1

PTSD child partial T1

ycp2

PTSD child partial T2

ycp3

PTSD child partial T3

cbin1

CBCL Internalizing T1

cbin3

CBCL Internalizing T3

cbex1

CBCL Externalizing T1

cbex3

CBCL Externalizing T3

bir1

Birlison T1

bir2

Birlison T2

bir3

Birlison T3

fdd.pred is the 65 by 65 binary predictor matrix used to impute fdd.

Details

Data from a randomized experiment to reduce post-traumatic stress by two treatments: Eye Movement Desensitization and Reprocessing (EMDR) (experimental treatment), and cognitive behavioral therapy (CBT) (control treatment). 52 children were randomized to one of these two treatments. Outcomes were measured at three time points: at baseline (pre-treatment, T1), post-treatment (T2, 4-8 weeks), and at follow-up (T3, 3 months). For more details, see de Roos et al (2011). Some person covariates were reshuffled. The imputation methodology is explained in Chapter 9 of van Buuren (2012).

Source

de Roos, C., Greenwald, R., den Hollander-Gijsman, M., Noorthoorn, E., van Buuren, S., de Jong, A. (2011). A Randomised Comparison of Cognitive Behavioral Therapy (CBT) and Eye Movement Desensitisation and Reprocessing (EMDR) in disaster-exposed children. European Journal of Psychotraumatology, 2, 5694.

Van Buuren, S. (2018). Flexible Imputation of Missing Data. Second Edition. Chapman & Hall/CRC. Boca Raton, FL. Boca Raton, FL.: Chapman & Hall/CRC Press.

Examples

data <- fdd
md.pattern(fdd)

[Package mice version 3.14.0 Index]