mads {mice}	R Documentation

Multivariate amputed data set (mads)

Description

The mads object is an S3 class that contains an amputed dataset, i.e., a dataset with simulated missing values. The ampute() function produces a mads object. The mads() function is the S3 constructor. The S3 class mads has the following methods: bwplot.mads(),print(), summary(), and xyplot.mads().

Usage

mads(
  call,
  prop,
  patterns,
  freq,
  mech,
  weights,
  cont,
  type,
  odds,
  amp,
  cand,
  scores,
  data
)

## S3 method for class 'mads'
print(x, ...)

## S3 method for class 'mads'
summary(object, ...)

Arguments

call
The function call that created the object.

prop
A numeric vector of proportions.

patterns
A matrix of missing data patterns.

freq
A numeric vector of frequencies for each pattern.

mech
A character string describing the missing data mechanism.

weights
A matrix of weights.

cont
A logical vector indicating continuous variables.

type
A character vector describing variable types.

odds
A matrix of odds ratios.

amp
A data frame for amplitude information.

cand
An integer vector of candidate variables.

scores
A list of scores.

data
A data frame representing the original data.

x
Object of class mads

...
Other parameters

object
Object of class mads

Value

An object of class "mads".

print() returns the input object invisibly.

summary() returns the input object invisibly.

Contents

call:

The function call.

prop:

Proportion of cases with missing values. Note: even when the proportion is entered as the proportion of missing cells (when bycases == TRUE), this object contains the proportion of missing cases.

patterns:

A data frame of size #patterns by #variables where 0 indicates a variable has missing values and 1 indicates a variable remains complete.

freq:

A vector of length #patterns containing the relative frequency with which the patterns occur. For example, if the vector is c(0.4, 0.4, 0.2), this means that of all cases with missing values, 40 percent is candidate for pattern 1, 40 percent for pattern 2 and 20 percent for pattern 3. The vector sums to 1.

mech:

A string specifying the missingness mechanism, either "MCAR", "MAR" or "MNAR".

weights:

A data frame of size #patterns by #variables. It contains the weights that were used to calculate the weighted sum scores. The weights may differ between patterns and between variables.

cont:

Logical, whether probabilities are based on continuous logit functions or on discrete odds distributions.

type:

A vector of strings containing the type of missingness for each pattern. Either "LEFT", "MID", "TAIL" or "RIGHT". The first type refers to the first pattern, the second type to the second pattern, etc.

odds:

A matrix where #patterns defines the #rows. Each row contains the odds of being missing for the corresponding pattern. The amount of odds values defines in how many quantiles the sum scores were divided. The values are relative probabilities: a quantile with odds value 4 will have a probability of being missing that is four times higher than a quantile with odds 1. The #quantiles may differ between patterns, NA is used for cells remaining empty.

amp:

A data frame containing the input data with NAs for the amputed values.

cand:

A vector that contains the pattern number for each case. A value between 1 and #patterns is given. For example, a case with value 2 is candidate for missing data pattern 2.

scores:

A list containing vectors with weighted sum scores of the candidates. The first vector refers to the candidates of the first pattern, the second vector refers to the candidates of the second pattern, etc. The length of the vectors differ because the number of candidates is different for each pattern.

data:

The complete data set that was entered in ampute.

Author(s)

Rianne Schouten, 2016

See Also

ampute, Vignette titled "Multivariate Amputation using Ampute".

[Package mice version 3.19.0 Index]