make.blots {mice}	R Documentation

Creates a blots argument

Description

This helper function creates a valid blots object. The blots object is an argument to the mice function. The name blots is a contraction of blocks-dots. Through blots, the user can specify any additional arguments that are specifically passed down to the lowest level imputation function.

Usage

make.blots(data, blocks = make.blocks(data))

Arguments

data
A data.frame with the source data

blocks
An optional specification for blocks of variables in the rows. The default assigns each variable in its own block.

Value

A matrix

See Also

make.blocks

Examples

make.predictorMatrix(nhanes)
make.blots(nhanes, blocks = name.blocks(c("age", "hyp"), "xxx"))

[Package mice version 3.14.0 Index]