make.post {mice}	R Documentation

Creates a post argument

Description

This helper function creates a valid post vector. The post vector is an argument to the mice function that specifies post-processing for a variable after each iteration of imputation.

Usage

make.post(data)

Arguments

data
A data frame or a matrix containing the incomplete data. Missing values are coded as NA.

Value

Character vector of ncol(data) element

See Also

mice

Examples

make.post(nhanes2)

[Package mice version 3.14.0 Index]