Creates a predictorMatrix argument

This helper function creates a valid predictMatrix. The predictorMatrix is an argument to the mice function. It specifies the target variable or block in the rows, and the predictor variables on the columns. An entry of 0 means that the column variable is NOT used to impute the row variable or block. A nonzero value indicates that it is used.

make.predictorMatrix(data, blocks = make.blocks(data))

Arguments

data
A data.frame with the source data

blocks
An optional specification for blocks of variables in the rows. The default assigns each variable in its own block.

Value

A matrix

See also

Examples

make.predictorMatrix(nhanes)

#>     age bmi hyp chl
#> age   0   1   1   1
#> bmi   1   0   1   1
#> hyp   1   1   0   1
#> chl   1   1   1   0

make.predictorMatrix(nhanes, blocks = make.blocks(nhanes, "collect"))

#>         age bmi hyp chl
#> collect   1   1   1   1