Imputation by logistic regression using the bootstrap

Imputes univariate missing data using logistic regression by a bootstrapped logistic regression model. The bootstrap method draws a simple bootstrap sample with replacement from the observed data y[ry] and x[ry, ].

mice.impute.logreg.boot(y, ry, x, wy = NULL, ...)

Arguments

y
Vector to be imputed

ry
Logical vector of length length(y) indicating the the subset y[ry] of elements in y to which the imputation model is fitted. The ry generally distinguishes the observed (TRUE) and missing values (FALSE) in y.

x
Numeric design matrix with length(y) rows with predictors for y. Matrix x may have no missing values.

wy
Logical vector of length length(y). A TRUE value indicates locations in y for which imputations are created.

...
Other named arguments.

Value

Vector with imputed data, same type as y, and of length sum(wy)

References

Van Buuren, S., Groothuis-Oudshoorn, K. (2011). mice: Multivariate Imputation by Chained Equations in R. Journal of Statistical Software, 45(3), 1-67. https://www.jstatsoft.org/v45/i03/

Van Buuren, S. (2018). Flexible Imputation of Missing Data. Second Edition. Chapman & Hall/CRC. Boca Raton, FL.

See also