Passive imputation

Calculate new variable during imputation

Usage

mice.impute.passive(data, func)

Arguments

data

A data frame

func

A formula specifying the transformations on data

Value

The result of applying formula

Details

Passive imputation is a special internal imputation function. Using this facility, the user can specify, at any point in the mice Gibbs sampling algorithm, a function on the imputed data. This is useful, for example, to compute a cubic version of a variable, a transformation like Q = W/H^2 based on two variables, or a mean variable like (x_1+x_2+x_3)/3. The so derived variables might be used in other places in the imputation model. The function allows to dynamically derive virtually any function of the imputed data at virtually any time.

References

Van Buuren, S., Groothuis-Oudshoorn, K. (2011). mice: Multivariate Imputation by Chained Equations in R. Journal of Statistical Software, 45(3), 1-67. doi:10.18637/jss.v045.i03

See also

Author

Stef van Buuren, Karin Groothuis-Oudshoorn, 2000