Imputation by the random indicator method for nonignorable data

Imputes nonignorable missing data by the random indicator method.

mice.impute.ri(y, ry, x, wy = NULL, ri.maxit = 10, ...)

Arguments

y
Vector to be imputed

ry
Logical vector of length length(y) indicating the the subset y[ry] of elements in y to which the imputation model is fitted. The ry generally distinguishes the observed (TRUE) and missing values (FALSE) in y.

x
Numeric design matrix with length(y) rows with predictors for y. Matrix x may have no missing values.

wy
Logical vector of length length(y). A TRUE value indicates locations in y for which imputations are created.

ri.maxit
Number of inner iterations

...
Other named arguments.

Value

Vector with imputed data, same type as y, and of length sum(wy)

Details

The random indicator method estimates an offset between the distribution of the observed and missing data using an algorithm that iterates over the response and imputation models.

This routine assumes that the response model and imputation model

References

Jolani, S. (2012). Dual Imputation Strategies for Analyzing Incomplete Data. Dissertation. University of Utrecht, Dec 7 2012.

See also