mice.theme {mice}	R Documentation

Set the theme for the plotting Trellis functions

Description

The mice.theme() function sets default choices for Trellis plots that are built into mice.

Usage

mice.theme(transparent = TRUE, alpha.fill = 0.3)

Arguments

transparent
A logical indicating whether alpha-transparency is allowed. The default is TRUE.

alpha.fill
A numerical values between 0 and 1 that indicates the default alpha value for fills.

Value

mice.theme() returns a named list that can be used as a theme in the functions in lattice. By default, the mice.theme() function sets transparent <- TRUE if the current device .Device supports semi-transparent colors.

Author(s)

Stef van Buuren 2011

[Package mice version 3.14.0 Index]