mira {mice}	R Documentation

Create an object of class "mira"

Description

The mira() functions constructs an S3 object representing a set of multiply imputed repeated analyses (mira). The default workflow generates the mira object using the with() function.

Usage

mira(
  call = match.call(),
  call1 = match.call(),
  nmis = integer(),
  analyses = list()
)

Arguments

call
The function call that created the object.

call1
A secondary function call, typically from the first imputation.

nmis
An integer vector representing the number of missing values.

analyses
A list of analyses performed on the imputed datasets.

Details

The as.mira() function takes the results of repeated complete-data analysis stored as a list, and turns it into a mira object that can be pooled.

In versions prior to mice 3.0 pooling required only that coef() and vcov() methods were available for fitted objects. This feature is no longer supported. The reason is that vcov() methods are inconsistent across packages, leading to buggy behaviour of the pool() function. Since mice 3.0+, the broom package takes care of filtering out the relevant parts of the complete-data analysis. It may happen that you'll see the messages like No method for tidying an S3 object of class ... or Error: No glance method for objects of class .... The royal way to solve this problem is to write your own glance() and tidy() methods and add these to broom according to the specifications given in https://broom.tidymodels.org.

The mira class of objects has methods for the following generic functions: print, summary.

Value

An object of class "mira". The mira class contains the following elements:

.Data:

Object of class "list" containing the following slots:

call:

The call that created the object.

call1:

The call that created the mids object that was used in call.

nmis:

An array containing the number of missing observations per column.

analyses:

A list of m components containing the individual fit objects from each of the m complete data analyses.

Author(s)

Stef van Buuren, Karin Groothuis-Oudshoorn, 2000

References

van Buuren S and Groothuis-Oudshoorn K (2011). mice: Multivariate Imputation by Chained Equations in R. Journal of Statistical Software, 45(3), 1-67. doi: 10.18637/jss.v045.i03

See Also

with.mids, mids, mipo

[Package mice version 3.19.0 Index]