Multiply imputed repeated analyses (mira)

The mira object is generated by the with.mids() function. The as.mira() function takes the results of repeated complete-data analysis stored as a list, and turns it into a mira object that can be pooled. Pooling requires that coef() and vcov() methods are available for fitted object. The mira class of objects has methods for the following generic functions: print, summary.

Note

Many of the functions of the mice package do not use the S4 class definitions, and instead rely on the S3 list equivalent oldClass(obj) <- "mira".

Slots

#'
.Data:

Object of class "list" containing the following slots:

call:

The call that created the object.

call1:

The call that created the mids object that was used in call.

nmis:

An array containing the number of missing observations per column.

analyses:

A list of m components containing the individual fit objects from each of the m complete data analyses.

References

van Buuren S and Groothuis-Oudshoorn K (2011). mice: Multivariate Imputation by Chained Equations in R. Journal of Statistical Software, 45(3), 1-67. https://www.jstatsoft.org/v45/i03/

See also