Name imputation blocks

This helper function names any unnamed elements in the blocks specification. This is a convenience function.

name.blocks(blocks, prefix = "B")

Arguments

blocks
List of vectors with variable names per block. List elements may be named to identify blocks. Variables within a block are imputed by a multivariate imputation method (see method argument). By default each variable is placed into its own block, which is effectively fully conditional specification (FCS) by univariate models (variable-by-variable imputation). Only variables whose names appear in blocks are imputed. The relevant columns in the where matrix are set to FALSE of variables that are not block members. A variable may appear in multiple blocks. In that case, it is effectively re-imputed each time that it is visited.

prefix
A character vector of length 1 with the prefix to be using for naming any unnamed blocks with two or more variables.

Value

A named list of character vectors with variables names.

Details

This function will name any unnamed list elements specified in the optional argument blocks. Unnamed blocks consisting of just one variable will be named after this variable. Unnamed blocks containing more than one variables will be named by the prefix argument, padded by an integer sequence stating at 1.

See also

Examples

blocks <- list(c("hyp", "chl"), AGE = "age", c("bmi", "hyp"), "edu")
name.blocks(blocks)

#> $B1
#> [1] "hyp" "chl"
#> 
#> $AGE
#> [1] "age"
#> 
#> $B2
#> [1] "bmi" "hyp"
#> 
#> $edu
#> [1] "edu"
#>