Hox pupil popularity data with missing popularity scores

Hox pupil popularity data with some missing popularity scores

Format

A data frame with 2000 rows and 7 columns:

pupil

Pupil number within school

school

School number

popular

Pupil popularity with 848 missing entries

sex

Pupil gender

texp

Teacher experience (years)

const

Constant intercept term

teachpop

Teacher popularity

Source

Hox, J. J. (2002) Multilevel analysis. Techniques and applications. Mahwah, NJ: Lawrence Erlbaum.

Details

The original, complete dataset was generated by Joop Hox as an example of well-behaved multilevel data set. The distributed data contains missing data in pupil popularity.

Examples

popmis[1:3,]

#>   pupil school popular sex texp const teachpop
#> 1     1      1      NA   1   24     1        7
#> 2     2      1      NA   0   24     1        7
#> 3     3      1       7   1   24     1        6