Summary of a mira object

Summary of a mira object

Print a mice.anova object

Usage

# S3 method for class 'mira'
summary(object, type = c("tidy", "glance", "summary"), ...)

# S3 method for class 'mice.anova'
summary(object, ...)

Arguments

object

A mira object

type

A length-1 character vector indicating the type of summary. There are three choices: type = "tidy" return the parameters estimates of each analyses as a data frame. type = "glance" return the fit statistics of each analysis as a data frame. type = "summary" returns a list of length m with the analysis results. The default is "tidy".

...

Other parameters passed down to print() and summary()

Value

NULL

NULL

See also