supports.transparent {mice}	R Documentation

Supports semi-transparent foreground colors?

Description

This function is used by mdc() to find out whether the current device supports semi-transparent foreground colors.

Usage

supports.transparent()

Details

The function calls the function dev.capabilities() from the package grDevices. The function return FALSE if the status of the current device is unknown.

Value

TRUE or FALSE

See Also

mdc dev.capabilities

Examples

supports.transparent()

[Package mice version 3.14.0 Index]