ampute.continuous {mice}	R Documentation

Multivariate amputation based on continuous probability functions

Description

This function creates a missing data indicator for each pattern. The continuous probability distributions (Van Buuren, 2012, pp. 63, 64) will be induced on the weighted sum scores, calculated earlier in the multivariate amputation function ampute.

Usage

ampute.continuous(P, scores, prop, type)

Arguments

P
A vector containing the pattern numbers of the cases's candidacies. For each case, a value between 1 and #patterns is given. For example, a case with value 2 is candidate for missing data pattern 2.

scores
A list containing vectors with the candidates's weighted sum scores, the result of an underlying function in ampute.

prop
A scalar specifying the proportion of missingness. Should be a value between 0 and 1. Default is a missingness proportion of 0.5.

type
A vector of strings containing the type of missingness for each pattern. Either "LEFT", "MID", "TAIL" or '"RIGHT". If a single missingness type is entered, all patterns will be created by the same type. If missingness types should differ over patterns, a vector of missingness types should be entered. Default is RIGHT for all patterns and is the result of ampute.default.type.

Value

A list containing vectors with 0 if a case should be made missing and 1 if a case should remain complete. The first vector refers to the first pattern, the second vector to the second pattern, etcetera.

Author(s)

Rianne Schouten [aut, cre], Gerko Vink [aut], Peter Lugtig [ctb], 2016

References

#'Van Buuren, S. (2018). Flexible Imputation of Missing Data. Second Edition. Chapman & Hall/CRC. Boca Raton, FL.

See Also

ampute, ampute.default.type

[Package mice version 3.14.0 Index]