ampute.default.odds {mice}	R Documentation

Default odds in ampute()

Description

Defines the default odds matrix for the multivariate amputation function ampute.

Usage

ampute.default.odds(patterns)

Arguments

patterns
A matrix of size #patterns by #variables where 0 indicates a variable should have missing values and 1 indicates a variable should remain complete. Could be the result of ampute.default.patterns.

Value

A matrix where #rows equals #patterns. Default is 4 quantiles with odds values 1, 2, 3 and 4, for each pattern, imitating a RIGHT type of missingness.

Author(s)

Rianne Schouten, 2016

See Also

ampute, ampute.default.patterns

[Package mice version 3.14.0 Index]