as.mira {mice}	R Documentation

Create a mira object from repeated analyses

Description

The as.mira() function takes the results of repeated complete-data analysis stored as a list, and turns it into a mira object that can be pooled.

Usage

as.mira(fitlist)

Arguments

fitlist
A list containing $m$ fitted analysis objects

Value

An S3 object of class mira.

Author(s)

Stef van Buuren

See Also

mira

[Package mice version 3.14.0 Index]