brandsma {mice}	R Documentation

Brandsma school data used Snijders and Bosker (2012)

Description

Dataset with raw data from Snijders and Bosker (2012) containing data from 4106 pupils attending 216 schools. This dataset includes all pupils and schools with missing data.

Format

brandsma is a data frame with 4106 rows and 14 columns:

sch

School number

pup

Pupil ID

iqv

IQ verbal

iqp

IQ performal

sex

Sex of pupil

ses

SES score of pupil

min

Minority member 0/1

rpg

Number of repeated groups, 0, 1, 2

lpr

language score PRE

lpo

language score POST

apr

Arithmetic score PRE

apo

Arithmetic score POST

den

Denomination classification 1-4 - at school level

ssi

School SES indicator - at school level

Note

This dataset is constructed from the raw data. There are a few differences with the data set used in Chapter 4 and 5 of Snijders and Bosker:

All schools are included, including the five school with missing values on langpost.

Missing denomina codes are left as missing.

Aggregates are undefined in the presence of missing data in the underlying values. Variables ses, iqv and iqp are in their original scale, and not globally centered. No aggregate variables at the school level are included.

There is a wider selection of original variables. Note however that the source data contain an even wider set of variables.

Source

Constructed from MLbook_2nded_total_4106-99.sav from https://www.stats.ox.ac.uk/~snijders/mlbook.htm by function data-raw/R/brandsma.R

References

Brandsma, HP and Knuver, JWM (1989), Effects of school and classroom characteristics on pupil progress in language and arithmetic. International Journal of Educational Research, 13(7), 777 - 788.

Snijders, TAB and Bosker RJ (2012). Multilevel Analysis, 2nd Ed. Sage, Los Angeles, 2012.

[Package mice version 3.14.0 Index]