bwplot.mads {mice}	R Documentation

Box-and-whisker plot of amputed and non-amputed data

Description

Plotting method to investigate the relation between the data variables and the amputed data. The function shows how the amputed values are related to the variable values.

Usage

## S3 method for class 'mads'
bwplot(
  x,
  data,
  which.pat = NULL,
  standardized = TRUE,
  descriptives = TRUE,
  layout = NULL,
  ...
)

Arguments

x
A mads (mads-class) object, typically created by ampute.

data
A string or vector of variable names that needs to be plotted. As a default, all variables will be plotted.

which.pat
A scalar or vector indicating which patterns need to be plotted. As a default, all patterns are plotted.

standardized
Logical. Whether the box-and-whisker plots need to be created from standardized data or not. Default is TRUE.

descriptives
Logical. Whether the mean, variance and n of the variables need to be printed. This is useful to examine the effect of the amputation. Default is TRUE.

layout
A vector of two values indicating how the boxplots of one pattern should be divided over the plot. For example, c(2, 3) indicates that the boxplots of six variables need to be placed on 3 rows and 2 columns. Default is 1 row and an amount of columns equal to #variables. Note that for more than 6 variables, multiple plots will be created automatically.

...
Not used, but for consistency with generic

Value

A list containing the box-and-whisker plots. Note that a new pattern will always be shown in a new plot.

Note

The mads object contains all the information you need to make any desired plots. Check mads-class or the vignette Multivariate Amputation using Ampute to understand the contents of class object mads.

Author(s)

Rianne Schouten, 2016

See Also

ampute, bwplot, Lattice for an overview of the package, mads-class

[Package mice version 3.14.0 Index]