Extract list of fitted models

Function getfit() returns the list of objects containing the repeated analysis results, or optionally, one of these fitted objects. The function looks for a list element called analyses, and return this component as a list with mira class. If element analyses is not found in x, then it returns x as a mira object.

getfit(x, i = -1L, simplify = FALSE)

Arguments

x

An object of class mira, typically produced by a call to with().

i

An integer between 1 and x$m signalling the index of the repeated analysis. The default i= -1 return a list with all analyses.

simplify

Should the return value be unlisted?

Value

If i = -1 an object of class mira containing all analyses. If i selects one of the analyses, then it return an object whose with class inherited from that element.

Details

No checking is done for validity of objects. The function also processes objects of class mitml.result from the mitml package.

See also

Author

Stef van Buuren, 2012, 2020

Examples

imp <- mice(nhanes, print = FALSE, seed = 21443)
fit <- with(imp, lm(bmi ~ chl + hyp))
f1 <- getfit(fit)
class(f1)
#> [1] "mira" "list"
f2 <- getfit(fit, 2)
class(f2)
#> [1] "lm"