Extract list of fitted model

getfit returns the list of objects containing the repeated analysis results, or optionally, one of these fit objects.

getfit(x, i = -1L, simplify = FALSE)

Arguments

x
An object of class mira or mitml.result, typically produced by a call to with().

i
An integer between 1 and x$m signaling the number of the repeated analysis. The default i= -1 return a list with all analyses.

simplify
Should the return value be unlisted?

Value

If i = -1 an object of class mitml.result containing all analyses, otherwise it returns the fitted object of the i'th repeated analysis.

See also

Examples

imp <- mice(nhanes)

#> 
#>  iter imp variable
#>   1   1  bmi  hyp  chl
#>   1   2  bmi  hyp  chl
#>   1   3  bmi  hyp  chl
#>   1   4  bmi  hyp  chl
#>   1   5  bmi  hyp  chl
#>   2   1  bmi  hyp  chl
#>   2   2  bmi  hyp  chl
#>   2   3  bmi  hyp  chl
#>   2   4  bmi  hyp  chl
#>   2   5  bmi  hyp  chl
#>   3   1  bmi  hyp  chl
#>   3   2  bmi  hyp  chl
#>   3   3  bmi  hyp  chl
#>   3   4  bmi  hyp  chl
#>   3   5  bmi  hyp  chl
#>   4   1  bmi  hyp  chl
#>   4   2  bmi  hyp  chl
#>   4   3  bmi  hyp  chl
#>   4   4  bmi  hyp  chl
#>   4   5  bmi  hyp  chl
#>   5   1  bmi  hyp  chl
#>   5   2  bmi  hyp  chl
#>   5   3  bmi  hyp  chl
#>   5   4  bmi  hyp  chl
#>   5   5  bmi  hyp  chl

fit <- with(imp, lm(bmi~chl+hyp))
getfit(fit)

#> Component 1 :
#> 
#> Call:
#> lm(formula = bmi ~ chl + hyp)
#> 
#> Coefficients:
#> (Intercept)          chl          hyp  
#>    20.23751      0.03661     -0.61411  
#> 
#> 
#> Component 2 :
#> 
#> Call:
#> lm(formula = bmi ~ chl + hyp)
#> 
#> Coefficients:
#> (Intercept)          chl          hyp  
#>    23.50501      0.01447      0.62077  
#> 
#> 
#> Component 3 :
#> 
#> Call:
#> lm(formula = bmi ~ chl + hyp)
#> 
#> Coefficients:
#> (Intercept)          chl          hyp  
#>    22.13509      0.02317      0.42998  
#> 
#> 
#> Component 4 :
#> 
#> Call:
#> lm(formula = bmi ~ chl + hyp)
#> 
#> Coefficients:
#> (Intercept)          chl          hyp  
#>    22.63828      0.02577     -0.73286  
#> 
#> 
#> Component 5 :
#> 
#> Call:
#> lm(formula = bmi ~ chl + hyp)
#> 
#> Coefficients:
#> (Intercept)          chl          hyp  
#>    25.75785      0.01223     -1.42520  
#> 
#> 

getfit(fit, 2)

#> 
#> Call:
#> lm(formula = bmi ~ chl + hyp)
#> 
#> Coefficients:
#> (Intercept)          chl          hyp  
#>    23.50501      0.01447      0.62077  
#>