glm.mids {mice}	R Documentation

Generalized linear model for mids object

Description

Applies glm() to a multiply imputed data set

Usage

glm.mids(formula, family = gaussian, data, ...)

Arguments

formula
a formula expression as for other regression models, of the form response ~ predictors. See the documentation of lm and formula for details.

family
The family of the glm model

data
An object of type mids, which stands for 'multiply imputed data set', typically created by function mice().

...
Additional parameters passed to glm.

Details

This function is included for backward compatibility with V1.0. The function is superseded by with.mids.

Value

An objects of class mira, which stands for 'multiply imputed repeated analysis'. This object contains data$m distinct glm.objects, plus some descriptive information.

Author(s)

Stef van Buuren, Karin Groothuis-Oudshoorn, 2000

References

Van Buuren, S., Groothuis-Oudshoorn, C.G.M. (2000) Multivariate Imputation by Chained Equations: MICE V1.0 User's manual. Leiden: TNO Quality of Life.

See Also

with.mids, glm, mids, mira

Examples

imp <- mice(nhanes)

# logistic regression on the imputed data
fit <- glm.mids((hyp == 2) ~ bmi + chl, data = imp, family = binomial)
fit

[Package mice version 3.14.0 Index]