Generalized linear model for mids object

Applies glm() to a multiply imputed data set

glm.mids(formula, family = gaussian, data, ...)

Arguments

formula
a formula expression as for other regression models, of the form response ~ predictors. See the documentation of lm and formula for details.

family
The family of the glm model

data
An object of type mids, which stands for 'multiply imputed data set', typically created by function mice().

…
Additional parameters passed to glm.

Value

An objects of class mira, which stands for 'multiply imputed repeated analysis'. This object contains data$m distinct glm.objects, plus some descriptive information.

Details

This function is included for backward compatibility with V1.0. The function is superseded by with.mids.

References

Van Buuren, S., Groothuis-Oudshoorn, C.G.M. (2000) Multivariate Imputation by Chained Equations: MICE V1.0 User's manual. Leiden: TNO Quality of Life.

See also

Examples

imp <- mice(nhanes)

#> 
#>  iter imp variable
#>   1   1  bmi  hyp  chl
#>   1   2  bmi  hyp  chl
#>   1   3  bmi  hyp  chl
#>   1   4  bmi  hyp  chl
#>   1   5  bmi  hyp  chl
#>   2   1  bmi  hyp  chl
#>   2   2  bmi  hyp  chl
#>   2   3  bmi  hyp  chl
#>   2   4  bmi  hyp  chl
#>   2   5  bmi  hyp  chl
#>   3   1  bmi  hyp  chl
#>   3   2  bmi  hyp  chl
#>   3   3  bmi  hyp  chl
#>   3   4  bmi  hyp  chl
#>   3   5  bmi  hyp  chl
#>   4   1  bmi  hyp  chl
#>   4   2  bmi  hyp  chl
#>   4   3  bmi  hyp  chl
#>   4   4  bmi  hyp  chl
#>   4   5  bmi  hyp  chl
#>   5   1  bmi  hyp  chl
#>   5   2  bmi  hyp  chl
#>   5   3  bmi  hyp  chl
#>   5   4  bmi  hyp  chl
#>   5   5  bmi  hyp  chl

# logistic regression on the imputed data
fit <- glm.mids((hyp==2)~bmi+chl, data=imp, family = binomial)
fit

#> call :
#> glm.mids(formula = (hyp == 2) ~ bmi + chl, family = binomial, 
#>     data = imp)
#> 
#> call1 :
#> mice(data = nhanes)
#> 
#> nmis :
#> age bmi hyp chl 
#>   0   9   8  10 
#> 
#> analyses :
#> [[1]]
#> 
#> Call:  glm(formula = formula, family = family, data = complete(data, 
#>     i))
#> 
#> Coefficients:
#> (Intercept)          bmi          chl  
#>    -6.37527      0.09658      0.01438  
#> 
#> Degrees of Freedom: 24 Total (i.e. Null);  22 Residual
#> Null Deviance:	    29.65 
#> Residual Deviance: 27.33 	AIC: 33.33
#> 
#> [[2]]
#> 
#> Call:  glm(formula = formula, family = family, data = complete(data, 
#>     i))
#> 
#> Coefficients:
#> (Intercept)          bmi          chl  
#>    -7.65351      0.13819      0.01414  
#> 
#> Degrees of Freedom: 24 Total (i.e. Null);  22 Residual
#> Null Deviance:	    29.65 
#> Residual Deviance: 26.71 	AIC: 32.71
#> 
#> [[3]]
#> 
#> Call:  glm(formula = formula, family = family, data = complete(data, 
#>     i))
#> 
#> Coefficients:
#> (Intercept)          bmi          chl  
#>    -6.89658      0.08831      0.01909  
#> 
#> Degrees of Freedom: 24 Total (i.e. Null);  22 Residual
#> Null Deviance:	    31.34 
#> Residual Deviance: 28.16 	AIC: 34.16
#> 
#> [[4]]
#> 
#> Call:  glm(formula = formula, family = family, data = complete(data, 
#>     i))
#> 
#> Coefficients:
#> (Intercept)          bmi          chl  
#>    -4.48685      0.02107      0.01334  
#> 
#> Degrees of Freedom: 24 Total (i.e. Null);  22 Residual
#> Null Deviance:	    25.02 
#> Residual Deviance: 23.98 	AIC: 29.98
#> 
#> [[5]]
#> 
#> Call:  glm(formula = formula, family = family, data = complete(data, 
#>     i))
#> 
#> Coefficients:
#> (Intercept)          bmi          chl  
#>     -5.1937      -0.1063       0.0337  
#> 
#> Degrees of Freedom: 24 Total (i.e. Null);  22 Residual
#> Null Deviance:	    27.55 
#> Residual Deviance: 21.53 	AIC: 27.53
#> 
#>