lm.mids {mice}	R Documentation

Linear regression for mids object

Description

Applies lm() to multiply imputed data set

Usage

lm.mids(formula, data, ...)

Arguments

formula
a formula object, with the response on the left of a ~ operator, and the terms, separated by + operators, on the right. See the documentation of lm and formula for details.

data
An object of type 'mids', which stands for 'multiply imputed data set', typically created by a call to function mice().

...
Additional parameters passed to lm

Details

This function is included for backward compatibility with V1.0. The function is superseded by with.mids.

Value

An objects of class mira, which stands for 'multiply imputed repeated analysis'. This object contains data$m distinct lm.objects, plus some descriptive information.

Author(s)

Stef van Buuren, Karin Groothuis-Oudshoorn, 2000

References

Van Buuren, S., Groothuis-Oudshoorn, K. (2011). mice: Multivariate Imputation by Chained Equations in R. Journal of Statistical Software, 45(3), 1-67. doi: 10.18637/jss.v045.i03

See Also

lm, mids, mira

Examples

imp <- mice(nhanes)
fit <- lm.mids(bmi ~ hyp + chl, data = imp)
fit

[Package mice version 3.14.0 Index]