make.formulas {mice}	R Documentation

Creates a formulas argument

Description

This helper function creates a valid formulas object. The formulas object is an argument to the mice function. It is a list of formula's that specifies the target variables and the predictors by means of the standard ~ operator.

Usage

make.formulas(data, blocks = make.blocks(data), predictorMatrix = NULL)

Arguments

data
A data.frame with the source data

blocks
An optional specification for blocks of variables in the rows. The default assigns each variable in its own block.

predictorMatrix
A predictorMatrix specified by the user.

Value

A list of formula's.

See Also

make.blocks, make.predictorMatrix

Examples

f1 <- make.formulas(nhanes)
f1
f2 <- make.formulas(nhanes, blocks = make.blocks(nhanes, "collect"))
f2

# for editing, it may be easier to work with the character vector
c1 <- as.character(f1)
c1

# fold it back into a formula list
f3 <- name.formulas(lapply(c1, as.formula))
f3

[Package mice version 3.14.0 Index]