Graphical parameter for missing data plots

mdc returns colors used to distinguish observed, missing and combined data in plotting. mice.theme return a partial list of named objects that can be used as a theme in stripplot, bwplot, densityplot and xyplot.

mdc(
  r = "observed",
  s = "symbol",
  transparent = TRUE,
  cso = grDevices::hcl(240, 100, 40, 0.7),
  csi = grDevices::hcl(0, 100, 40, 0.7),
  csc = "gray50",
  clo = grDevices::hcl(240, 100, 40, 0.8),
  cli = grDevices::hcl(0, 100, 40, 0.8),
  clc = "gray50"
)

Arguments

r
A numerical or character vector. The numbers 1-6 request colors as follows: 1=cso, 2=csi, 3=csc, 4=clo, 5=cli and 6=clc. Alternatively, r may contain the strings ' observed', 'missing', or 'both', or abbreviations thereof.

s
A character vector containing the strings 'symbol' or ' line', or abbreviations thereof.

transparent
A logical indicating whether alpha-transparency is allowed. The default is TRUE.

cso
The symbol color for the observed data. The default is a transparent blue.

csi
The symbol color for the missing or imputed data. The default is a transparent red.

csc
The symbol color for the combined observed and imputed data. The default is a grey color.

clo
The line color for the observed data. The default is a slightly darker transparent blue.

cli
The line color for the missing or imputed data. The default is a slightly darker transparent red.

clc
The line color for the combined observed and imputed data. The default is a grey color.

Value

mdc() returns a vector containing color definitions. The length of the output vector is calculate from the length of r and s. Elements of the input vectors are repeated if needed.

Details

This function eases consistent use of colors in plots. The default follows the Abayomi convention, which uses blue for observed data, red for missing or imputed data, and black for combined data.

References

Sarkar, Deepayan (2008) Lattice: Multivariate Data Visualization with R, Springer.

See also

Author

Stef van Buuren, sept 2012.

Examples

# all six colors
mdc(1:6)
#> [1] "#006CC2B3" "#B61A51B3" "gray50"    "#006CC2CC" "#B61A51CC" "gray50"   

# lines color for observed and missing data
mdc(c("obs", "mis"), "lin")
#> [1] "#006CC2CC" "#B61A51CC"