Multiply imputed data set (mids)

The mids object contains a multiply imputed data set. The mids object is generated by functions mice(), mice.mids(), cbind.mids(), rbind.mids() and ibind.mids().

Details

The mids class of objects has methods for the following generic functions: print, summary, plot.

The loggedEvents entry is a matrix with five columns containing a record of automatic removal actions. It is NULL is no action was made. At initialization the program does the following three actions:

1

A variable that contains missing values, that is not imputed and that is used as a predictor is removed

2

A constant variable is removed

3

A collinear variable is removed.

During iteration, the program does the following actions:

1

One or more variables that are linearly dependent are removed (for categorical data, a 'variable' corresponds to a dummy variable)

2

Proportional odds regression imputation that does not converge and is replaced by polyreg.

Explanation of elements in loggedEvents:

it

iteration number at which the record was added,

im

imputation number,

dep

name of the dependent variable,

meth

imputation method used,

out

a (possibly long) character vector with the names of the altered or removed predictors.

Note

The mice package does not use the S4 class definitions, and instead relies on the S3 list equivalent oldClass(obj) <- "mids".

Slots

.Data:

Object of class "list" containing the following slots:

data:

Original (incomplete) data set.

imp:

A list of ncol(data) components with the generated multiple imputations. Each list component is a data.frame (nmis[j] by m) of imputed values for variable j. A NULL component is used for variables for which not imputations are generated.

m:

Number of imputations.

where:

The where argument of the mice() function.

blocks:

The blocks argument of the mice() function.

call:

Call that created the object.

nmis:

An array containing the number of missing observations per column.

method:

A vector of strings of length(blocks specifying the imputation method per block.

predictorMatrix:

A numerical matrix of containing integers specifying the predictor set.

visitSequence:

A vector of variable and block names that specifies how variables and blocks are visited in one iteration throuh the data.

formulas:

A named list of formula's, or expressions that can be converted into formula's by as.formula. List elements correspond to blocks. The block to which the list element applies is identified by its name, so list names must correspond to block names.

post:

A vector of strings of length length(blocks) with commands for post-processing.

blots:

"Block dots". The blots argument to the mice() function.

ignore:

A logical vector of length nrow(data) indicating the rows in data used to build the imputation model. (new in mice 3.12.0)

seed:

The seed value of the solution.

iteration:

Last Gibbs sampling iteration number.

lastSeedValue:

The most recent seed value.

chainMean:

An array of dimensions ncol by maxit by m elements containing the mean of the generated multiple imputations. The array can be used for monitoring convergence. Note that observed data are not present in this mean.

chainVar:

An array with similar structure as chainMean, containing the variance of the imputed values.

loggedEvents:

A data.frame with five columns containing warnings, corrective actions, and other inside info.

version:

Version number of mice package that created the object.

date:

Date at which the object was created.

References

van Buuren S and Groothuis-Oudshoorn K (2011). mice: Multivariate Imputation by Chained Equations in R. Journal of Statistical Software, 45(3), 1-67. doi:10.18637/jss.v045.i03

See also

Author

Stef van Buuren, Karin Groothuis-Oudshoorn, 2000