mids2mplus {mice}	R Documentation

Export mids object to Mplus

Description

Converts a mids object into a format recognized by Mplus, and writes the data and the Mplus input files

Usage

mids2mplus(
  imp,
  file.prefix = "imp",
  path = getwd(),
  sep = "\t",
  dec = ".",
  silent = FALSE
)

Arguments

imp
The imp argument is an object of class mids, typically produced by the mice() function.

file.prefix
A character string describing the prefix of the output data files.

path
A character string containing the path of the output file. By default, files are written to the current R working directory.

sep
The separator between the data fields.

dec
The decimal separator for numerical data.

silent
A logical flag stating whether the names of the files should be printed.

Details

This function automates most of the work needed to export a mids object to Mplus. The function writes the multiple imputation datasets, the file that contains the names of the multiple imputation data sets and an Mplus input file. The Mplus input file has the proper file names, so in principle it should run and read the data without alteration. Mplus will recognize the data set as a multiply imputed data set, and do automatic pooling in procedures where that is supported.

Value

The return value is NULL.

Author(s)

Gerko Vink, 2011.

See Also

mids, mids2spss

[Package mice version 3.14.0 Index]