mipo {mice}	R Documentation

mipo: Multiple imputation pooled object

Description

The mipo object contains the results of the pooling step. The function pool generates an object of class mipo.

Usage

mipo(mira.obj, ...)

## S3 method for class 'mipo'
summary(
  object,
  type = c("tests", "all"),
  conf.int = FALSE,
  conf.level = 0.95,
  exponentiate = FALSE,
  ...
)

## S3 method for class 'mipo'
print(x, ...)

## S3 method for class 'mipo.summary'
print(x, ...)

process_mipo(z, x, conf.int = FALSE, conf.level = 0.95, exponentiate = FALSE)

Arguments

mira.obj
An object of class mira

...
Arguments passed down

object
An object of class mipo

conf.int
Logical indicating whether to include a confidence interval. The default is FALSE.

conf.level
Confidence level of the interval, used only if conf.int = TRUE. Number between 0 and 1.

exponentiate
Flag indicating whether to exponentiate the coefficient estimates and confidence intervals (typical for logistic regression).

x
An object of class mipo

z
Data frame with a tidied version of a coefficient matrix

Details

An object class mipo is a list with elements: call, m, pooled and glanced.

The pooled elements is a data frame with columns:

estimate	 Pooled complete data estimate

ubar 	 Within-imputation variance of estimate

b 	 Between-imputation variance of estimate

t 	 Total variance, of estimate

dfcom 	 Degrees of freedom in complete data

df 	 Degrees of freedom of $t$-statistic

riv 	 Relative increase in variance

lambda 	 Proportion attributable to the missingness

fmi 	 Fraction of missing information

The names of the terms are stored as row.names(pooled).

The glanced elements is a data.frame with m rows. The precise composition depends on the class of the complete-data analysis. At least field nobs is expected to be present.

The process_mipo is a helper function to process a tidied mipo object, and is normally not called directly. It adds a confidence interval, and optionally exponentiates, the result.

Value

The summary method returns a data frame with summary statistics of the pooled analysis.

References

van Buuren S and Groothuis-Oudshoorn K (2011). mice: Multivariate Imputation by Chained Equations in R. Journal of Statistical Software, 45(3), 1-67. doi: 10.18637/jss.v045.i03

See Also

pool, mids, mira

[Package mice version 3.14.0 Index]