norm.draw {mice}	R Documentation

Draws values of beta and sigma by Bayesian linear regression

Description

This function draws random values of beta and sigma under the Bayesian linear regression model as described in Rubin (1987, p. 167). This function can be called by user-specified imputation functions.

Usage

norm.draw(y, ry, x, rank.adjust = TRUE, ...)

.norm.draw(y, ry, x, rank.adjust = TRUE, ...)

Arguments

y
Incomplete data vector of length n

ry
Vector of missing data pattern (FALSE=missing, TRUE=observed)

x
Matrix (n x p) of complete covariates.

rank.adjust
Argument that specifies whether NA's in the coefficients need to be set to zero. Only relevant when ls.meth = "qr" AND the predictor matrix is rank-deficient.

...
Other named arguments.

Value

A list containing components coef (least squares estimate), beta (drawn regression weights) and sigma (drawn value of the residual standard deviation).

Author(s)

Gerko Vink, 2018, for this version, based on earlier versions written by Stef van Buuren, Karin Groothuis-Oudshoorn, 2017

References

Rubin, D.B. (1987). Multiple imputation for nonresponse in surveys. New York: Wiley.

[Package mice version 3.14.0 Index]