Pools R^2 of m models fitted to multiply-imputed data

The function pools the coefficients of determination R^2 or the adjusted coefficients of determination (R^2_a) obtained with the lm modeling function. For pooling it uses the Fisher z-transformation.

Usage

pool.r.squared(object, adjusted = FALSE)

Arguments

object

An object of class 'mira' or 'mipo', produced by lm.mids, with.mids, or pool with lm as modeling function.

adjusted

A logical value. If adjusted=TRUE then the adjusted R^2 is calculated. The default value is FALSE.

Value

Returns a 1x4 table with components. Component est is the pooled R^2 estimate. Component lo95 is the 95 \ Component hi95 is the 95 \ Component fmi is the fraction of missing information due to nonresponse.

References

Harel, O (2009). The estimation of R^2 and adjusted R^2 in incomplete data sets using multiple imputation, Journal of Applied Statistics, 36:1109-1118.

Rubin, D.B. (1987). Multiple Imputation for Nonresponse in Surveys. New York: John Wiley and Sons.

van Buuren S and Groothuis-Oudshoorn K (2011). mice: Multivariate Imputation by Chained Equations in R. Journal of Statistical Software, 45(3), 1-67. doi:10.18637/jss.v045.i03

See also

Author

Karin Groothuis-Oudshoorn and Stef van Buuren, 2009

Examples

imp <- mice(nhanes, print = FALSE, seed = 16117)
fit <- with(imp, lm(chl ~ age + hyp + bmi))

# input: mira object
pool.r.squared(fit)
#>           est      lo 95     hi 95       fmi
#> R^2 0.4338408 0.06503877 0.7513683 0.4176739
pool.r.squared(fit, adjusted = TRUE)
#>               est      lo 95     hi 95       fmi
#> adj R^2 0.3507643 0.01771032 0.7091501 0.4617807

# input: mipo object
est <- pool(fit)
pool.r.squared(est)
#>           est      lo 95     hi 95       fmi
#> R^2 0.4338408 0.06503877 0.7513683 0.4176739
pool.r.squared(est, adjusted = TRUE)
#>               est      lo 95     hi 95       fmi
#> adj R^2 0.3507643 0.01771032 0.7091501 0.4617807