Multiple imputation pooling: univariate version

Pools univariate estimates of m repeated complete data analysis

pool.scalar(Q, U, n = Inf, k = 1)

Arguments

Q
A vector of univariate estimates of m repeated complete data analyses.

U
A vector containing the corresponding m variances of the univariate estimates.

n
A number providing the sample size. If nothing is specified, an infinite sample n = Inf is assumed.

k
A number indicating the number of parameters to be estimated. By default, k = 1 is assumed.

Value

Returns a list with components. Component m is the number of imputations. Component qhat contains the m univariate estimates of repeated complete data analyses. Component u contains the corresponding m variances of the univariate estimates. Component qbar is the pooled univariate estimate, formula (3.1.2) Rubin (1987). Component ubar is the mean of the variances (i.e. the pooled within-imputation variance), formula (3.1.3) Rubin (1987). Component b is the between-imputation variance, formula (3.1.4) Rubin (1987). Component t is the total variance of the pooled estimated, formula (3.1.5) Rubin (1987). Component r is the relative increase in variance due to nonresponse, formula (3.1.7) Rubin (1987). Component df is the degrees of freedom for t reference distribution, formula (3.1.6) Rubin (1987) or method of Barnard-Rubin (1999) (if method = "smallsample"). Component fmi is the fraction missing information due to nonresponse, formula (3.1.10) Rubin (1987).

Details

The function averages the univariate estimates of the complete data model, computes the total variance over the repeated analyses, and computes the relative increase in variance due to nonresponse and the fraction of missing information.

References

Rubin, D.B. (1987). Multiple Imputation for Nonresponse in Surveys. New York: John Wiley and Sons.

See also

Examples

imp <- mice(nhanes)

#> 
#>  iter imp variable
#>   1   1  bmi  hyp  chl
#>   1   2  bmi  hyp  chl
#>   1   3  bmi  hyp  chl
#>   1   4  bmi  hyp  chl
#>   1   5  bmi  hyp  chl
#>   2   1  bmi  hyp  chl
#>   2   2  bmi  hyp  chl
#>   2   3  bmi  hyp  chl
#>   2   4  bmi  hyp  chl
#>   2   5  bmi  hyp  chl
#>   3   1  bmi  hyp  chl
#>   3   2  bmi  hyp  chl
#>   3   3  bmi  hyp  chl
#>   3   4  bmi  hyp  chl
#>   3   5  bmi  hyp  chl
#>   4   1  bmi  hyp  chl
#>   4   2  bmi  hyp  chl
#>   4   3  bmi  hyp  chl
#>   4   4  bmi  hyp  chl
#>   4   5  bmi  hyp  chl
#>   5   1  bmi  hyp  chl
#>   5   2  bmi  hyp  chl
#>   5   3  bmi  hyp  chl
#>   5   4  bmi  hyp  chl
#>   5   5  bmi  hyp  chl

m <- imp$m
Q <- rep(NA, m)
U <- rep(NA, m)
for (i in 1:m) {
   Q[i] <- mean(complete(imp, i)$bmi)
   U[i] <- var(complete(imp, i)$bmi) / nrow(nhanes)  # (standard error of estimate)^2
}
pool.scalar(Q, U, n = nrow(nhanes), k = 1)  # Barnard-Rubin 1999

#> $m
#> [1] 5
#> 
#> $qhat
#> [1] 26.564 26.832 27.412 26.540 26.532
#> 
#> $u
#> [1] 0.5947960 0.7992907 0.8649440 0.6654000 0.6145907
#> 
#> $qbar
#> [1] 26.776
#> 
#> $ubar
#> [1] 0.7078043
#> 
#> $b
#> [1] 0.141952
#> 
#> $t
#> [1] 0.8781467
#> 
#> $df
#> [1] 15.32876
#> 
#> $r
#> [1] 0.2406631
#> 
#> $fmi
#> [1] 0.2819309
#>