popmis {mice}	R Documentation

Hox pupil popularity data with missing popularity scores

Description

Hox pupil popularity data with some missing popularity scores

Format

A data frame with 2000 rows and 7 columns:

pupil

Pupil number within school

school

School number

popular

Pupil popularity with 848 missing entries

sex

Pupil gender

texp

Teacher experience (years)

const

Constant intercept term

teachpop

Teacher popularity

Details

The original, complete dataset was generated by Joop Hox as an example of well-behaved multilevel data set. The distributed data contains missing data in pupil popularity.

Source

Hox, J. J. (2002) Multilevel analysis. Techniques and applications. Mahwah, NJ: Lawrence Erlbaum.

Examples

popmis[1:3, ]

[Package mice version 3.14.0 Index]